+++ R/baggr-package.R
@@ -1,7 +1,7 @@
 #' baggr - a package for Bayesian meta-analysis
 #'
 #' This is _baggr_ (pronounced as _bagger_ or _badger_), a Bayesian meta-analysis
-#' package for R using [Stan](https://mc-stan.org/).
+#' package for R that uses [Stan](https://mc-stan.org/) to fit the models.
 #' _Baggr_ is intended to be user-friendly and transparent so that
 #' it's easier to understand the models you are building and criticise them.
 #'
@@ -12,7 +12,7 @@
 #' [random effects models](https://stats.stackexchange.com/questions/4700/what-is-the-difference-between-fixed-effect-random-effect-and-mixed-effect-mode))
 #' with weakly informative priors by calling [Stan](https://mc-stan.org/) to carry
 #' out Bayesian inference. Modelling of variances or quantiles, standardisation and
-#' transformation of data is also possible.
+#' transformation of data are also possible.
 #'
 #'
 #' @section Getting help:
@@ -30,5 +30,6 @@
 #' @import rstantools
 #' @importFrom rstan sampling
 #'
-
+#' @references
+#' Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
 NULL