+++ R/trt_effects.R
@@ -23,18 +23,18 @@ treatment_effect <- function(bg, summary = FALSE,
     message("There is no treatment effect estimated when pooling = 'none'.")
     return(list(tau = as.numeric(NA), sigma_tau = as.numeric(NA)))
   }
-  if(bg$model %in% c("rubin", "mutau", "logit", "full")) {
+  if(bg$model %in% c("rubin", "mutau", "mutau_full", "logit", "rubin_full")) {
     tau <- rstan::extract(bg$fit, pars="mu")[[1]]
     if(bg$model %in% c("rubin", "logit"))
       tau <- c(tau)
-    if(bg$model == "mutau")
+    if(bg$model %in% c("mutau", "mutau_full"))
       tau <- tau[,1,2]
 
     if(bg$pooling == "partial"){
       sigma_tau <- rstan::extract(bg$fit, pars="tau")[[1]]
       if(bg$model %in% c("rubin", "logit"))
         sigma_tau <- c(sigma_tau)
-      if(bg$model == "mutau")
+      if(bg$model %in% c("mutau", "mutau_full"))
         sigma_tau <- sqrt(sigma_tau[,1,2,2])
     }
     if(bg$pooling == "full")
@@ -46,11 +46,27 @@ treatment_effect <- function(bg, summary = FALSE,
     sigma_tau <- t(apply(rstan::extract(bg$fit, "Sigma_1")[[1]], 1, diag))
     # in model with correlation, we have Var(), not SD()
     sigma_tau <- sqrt(sigma_tau)
+  } else if(bg$model == "sslab") {
+    mean_params <- c("tau[1]", "tau[2]",
+                     "sigma_TE[1]", "sigma_TE[2]",
+                     "kappa[1,1,2]", "kappa[1,2,2]")
+    sigma_params <- c("hypersd_tau[1]", "hypersd_tau[2]",
+                      "hypersd_sigma_TE[1]", "hypersd_sigma_TE[2]",
+                      "hypersd_kappa[1,1,2]", "hypersd_kappa[1,2,2]")
+    tau <- as.matrix(bg$fit, mean_params)
+    sigma_tau <- as.matrix(bg$fit, sigma_params)
+
+  } else {
+    stop("Can't calculate treatment effect for this model.")
   }
+
+  if(length(bg$effects) > 1)
+    colnames(tau) <- colnames(sigma_tau) <- bg$effects
+
   if(!is.null(transform)){
     tau <- do.call(transform, list(tau))
     sigma_tau <- NA # by convention we set it to NA so that people don't transform
-                    # and then do operations on it by accident
+    # and then do operations on it by accident
   }
   if(summary) {
     tau <- mint(tau, int=interval, median=TRUE, sd = TRUE)
@@ -64,19 +80,20 @@ treatment_effect <- function(bg, summary = FALSE,
 
 #' Make predictive draws from baggr model
 #'
-#' This function takes the samples of hyperparameters of a [baggr] model
-#' (typically hypermean and hyper-SD returned by [treatment_effect]) and draws values of
+#' This function takes the samples of hyperparameters from a [baggr] model
+#' (typically hypermean and hyper-SD, which you can see using [treatment_effect]) and draws values of
 #' new realisations of treatment effect, i.e. an additional draw from the "population of studies".
 #' This can be used for both prior and posterior draws, depending on [baggr] model.
 #'
 #' @param x A `baggr` class object.
-#' @param transform a transformation to apply to the result, should be an R function;
-#'                  (this is commonly used when calling `group_effects` from other
-#'                  plotting or printing functions)
+#' @param transform a transformation (an R function) to apply to the result of a draw.
+#' @param summary logical; if TRUE returns summary statistics rather than samples from the distribution;
+#' @param interval uncertainty interval width (numeric between 0 and 1), if `summary=TRUE`
 #' @param n How many values to draw? The default is as long as the number of samples
 #'          in the `baggr` object (see _Details_).
 #'
-#' @return A vector of possible values of the treatment effect.
+#' @return A vector (with `n` values) for models with one treatment effect parameter,
+#'         a matrix (`n` rows and same number of columns as number of parameters) otherwise.
 #' @export
 #'
 #' @seealso [treatment_effect] returns samples of hypermean and hyper-SD
@@ -101,9 +118,9 @@ treatment_effect <- function(bg, summary = FALSE,
 #' @references
 #' Riley, Richard D., Julian P. T. Higgins, and Jonathan J. Deeks.
 #' "Interpretation of Random Effects Meta-Analyses".
-#' _BMJ 342 (10 February 2011)._ <https://doi.org/10.1136/bmj.d549>.
+#' _BMJ 342 (10 February 2011)._ \doi{10.1136/bmj.d549}.
 #'
-effect_draw <- function(x, n, transform=NULL) {
+effect_draw <- function(x, n, transform=NULL, summary = FALSE, interval = .95) {
   check_if_baggr(x)
 
   te <- treatment_effect(x)
@@ -114,7 +131,7 @@ effect_draw <- function(x, n, transform=NULL) {
     if(neffects > 1){
       if(n > nrow(te$tau))
         warning("Making more effect draws than there are available samples in Stan object.",
-                "Consider running baggr() with higher iter=.")
+                "Consider running baggr() with higher iter= setting")
       rows <- sample(nrow(te$tau), n, replace = TRUE)
       te$tau   <- te$tau[rows,]
       te$sigma_tau <- te$sigma_tau[rows,]
@@ -122,20 +139,26 @@ effect_draw <- function(x, n, transform=NULL) {
     if(neffects == 1){
       if(n > length(te$tau))
         warning("Making more effect draws than there are available samples in Stan object.",
-                "Consider running baggr() with higher iter=.")
+                "Consider running baggr() with higher iter= setting")
       rows <- sample(length(te$tau), n, replace = TRUE)
       te$tau   <- te$tau[rows]
       te$sigma_tau <- te$sigma_tau[rows]
     }
   }
 
-  # Make draws using normal distribution:
+  # Make draws using normal distribution (for now it's the only option)
   new_tau <- rnorm(length(te$tau), c(te$tau), c(te$sigma_tau))
-  if(neffects > 1)
-    new_tau <- matrix(new_tau, nrow(te$tau), ncol(te$tau))
 
+  if(neffects > 1){
+    new_tau <- matrix(new_tau, nrow(te$tau), ncol(te$tau))
+    colnames(new_tau) <- colnames(te$tau)
+  }
   if(!is.null(transform))
     new_tau <- do.call(transform, list(new_tau))
+
+  if(summary) {
+    new_tau <- mint(new_tau, int=interval, median=TRUE, sd = TRUE)
+  }
 
   new_tau
 }
@@ -149,7 +172,8 @@ effect_draw <- function(x, n, transform=NULL) {
 #' for one or more `baggr` objects.
 #'
 #' @param ... Object(s) of class [baggr]. If there is more than one,
-#'            the names of objects will be used as a plot legend (see example).
+#'            a comparison will be plotted and  names of objects
+#'            will be used as a plot legend (see examples).
 #' @param transform a transformation to apply to the result, should be an R function;
 #'                  (this is commonly used when calling `group_effects` from other
 #'                  plotting or printing functions)
@@ -196,7 +220,7 @@ effect_plot <- function(..., transform=NULL) {
   if(all(unlist(lapply(l, attr, "ppd"))))
     caption <- list(
       title = "Prior distribution for possible treatment effect",
-      subtitle = "No data, only sampling from prior"
+      subtitle = "No data used, only sampling from priors"
     )
   if(is.null(names(l))){
     if(length(l) > 1)
@@ -205,17 +229,36 @@ effect_plot <- function(..., transform=NULL) {
   }
 
   # Check effects and prepare X label
-  if(any(unlist(lapply(l, function(x) length(x$effects))) > 1))
-    stop("Effect_plot is only possible for models with 1-dimensional treatment effects")
-  effects <- paste("Treatment effect on", unique(unlist(lapply(l, function(x) x$effects))))
-  if(length(effects) > 1)
+  # if(any(unlist(lapply(l, function(x) length(x$effects))) > 1))
+  # stop("Effect_plot is only possible for models with 1-dimensional treatment effects")
+  # effects <- paste("Treatment effect on",
+  # unique(unlist(lapply(l, function(x) x$effects))))
+  effects <- unique(unlist(lapply(l, function(x) x$effects)))
+  n_parameters <- unique(unlist(lapply(l, function(x) x$n_parameters)))
+  if(length(n_parameters) != 1)
+    stop("All models must have the same number of parameters")
+  if(length(effects) > n_parameters)
     stop("All models must have same effects")
 
+
+
+  # dim <-
   l <- lapply(l, effect_draw, transform=transform)
+
   df <- data.frame()
-  for(i in seq_along(l))
-    df <- rbind(df, data.frame("model"=names(l)[i],
-                               "value" = l[[i]]))
+  for(i in seq_along(l)){
+    if(n_parameters == 1)
+      df <- rbind(df, data.frame("model"=names(l)[i],
+                                 "value" = l[[i]]))
+    else{
+      # Melt with base R
+      for(nm in colnames(l[[i]]))
+        df <- rbind(df, data.frame("model"=names(l)[i],
+                                   "variable" = nm,
+                                   "value" = l[[i]][,nm]))
+    }
+  }
+
   single_model_flag <- (length(l) == 1)
   model <- value <- NULL
   ggplot(df, aes(value, group = model, fill = model)) +
@@ -223,6 +266,7 @@ effect_plot <- function(..., transform=NULL) {
     geom_density(alpha = .25) +
     ggtitle(label = caption$title,
             subtitle = caption$subtitle) +
-    xlab(effects) +
-    {if(single_model_flag) theme(legend.position = "none")}
+    {if(n_parameters == 1) xlab(effects)} +
+    {if(single_model_flag) theme(legend.position = "none")} +
+    {if(n_parameters > 1) facet_wrap(~variable, ncol = 3, scales = "free")}
 }