+++ tests/testthat/test_full.R
@@ -1,4 +1,4 @@
-context("baggr() calls with full model")
+context("baggr() calls with IPD version of Rubin model")
 library(baggr)
 set.seed(1990)
 
@@ -6,6 +6,8 @@ set.seed(1990)
 schools_ipd <- data.frame()
 N <- c(rep(10, 4), rep(20, 4))
 for(i in 1:8){
+  bsl <- rnorm(1, 0, 5)
+
   x <- rnorm(N[i])
   x <- (x-mean(x))/sd(x)
   x <- x*schools$se[i]*sqrt(N[i])/1.41 + schools$tau[i]
@@ -15,9 +17,9 @@ for(i in 1:8){
   y <- y*schools$se[i]*sqrt(N[i])/1.41
 
   schools_ipd <- rbind(schools_ipd,
-                       data.frame(group = schools$group[i], outcome = x, treatment = 1),
+                       data.frame(group = schools$group[i], outcome = bsl + x, treatment = 1),
                        # This is just so that we don't trip off prepare_ma:
-                       data.frame(group = schools$group[i], outcome = y, treatment = 0))
+                       data.frame(group = schools$group[i], outcome = bsl + y, treatment = 0))
 }
 
 
@@ -26,13 +28,13 @@ bg_p <- expect_warning(baggr(schools_ipd, pooling = "partial", iter = 150, refre
 bg_f <- expect_warning(baggr(schools_ipd, pooling = "full", iter = 150, refresh=0))
 
 
-test_that("Different pooling methods work for the full model", {
+test_that("Different pooling methods work for the rubin_full model", {
   expect_is(bg_n, "baggr")
   expect_is(bg_p, "baggr")
   expect_is(bg_f, "baggr")
 })
 
-test_that("Basic operations on full data model", {
+test_that("Basic operations on rubin_full model", {
   expect_error(baggr(schools_ipd, rubbish = 41))
   expect_is(pooling(bg_p)[,,1], "matrix")
   expect_is(plot(bg_p), "gg")
@@ -42,7 +44,7 @@ test_that("Basic operations on full data model", {
   expect_is(bgc, "baggr_compare")
 })
 
-test_that("Full model crashes with nonsense inputs", {
+test_that("rubin_full model crashes with nonsense inputs", {
   expect_error(baggr(schools_ipd, outcome = 2), "Arguments")
   expect_error(baggr(schools_ipd, group = 2), "Arguments")
   expect_error(baggr(schools_ipd, treatment = 2), "Arguments")
@@ -52,6 +54,13 @@ test_that("Full model crashes with nonsense inputs", {
   schools_ipd2 <- schools_ipd; schools_ipd2$outcome <- as.character(schools_ipd$outcome)
   expect_error(baggr(schools_ipd2), "has to be numeric")
 
+})
+
+test_that("Using old syntax (model = full) still works", {
+  bg <- expect_warning(expect_message(baggr(schools_ipd,
+                                            model = "full",
+                                            pooling = "none", iter = 10, refresh=0)))
+  expect_is(bg, "baggr")
 })
 
 
@@ -65,7 +74,7 @@ comp_flpr <- expect_warning(baggr_compare(
 
 comp_flmdls <- baggr_compare(bg_f, bg_p)
 
-test_that("baggr comparison method works for Full model", {
+test_that("baggr comparison method works for rubin_full model", {
 
   expect_is(comp_flpl, "baggr_compare")
   expect_is(comp_flpr, "baggr_compare")
@@ -79,21 +88,32 @@ test_that("baggr comparison method works for Full model", {
   expect_gt(length(comp_flpr), 0)
   expect_gt(length(comp_flmdls), 0)
 
-  expect_is(plot(comp_flpl), "plot_list")
-  expect_is(plot(comp_flpl)[[1]], "ggplot")
+  expect_is(plot(comp_flpl), "gg")
 
   expect_is(plot(comp_flpr), "ggplot")
 
-  expect_is(plot(comp_flmdls), "plot_list")
-  expect_is(plot(comp_flmdls)[[1]], "ggplot")
+  expect_is(plot(comp_flmdls), "gg")
 
-  expect_is(plot(comp_flpl, arrange = "grid"), "plot_list")
-  expect_is(plot(comp_flpl, arrange = "grid")[[1]], "ggplot")
+  expect_is(plot(comp_flpl, grid_models = TRUE), "gtable")
 
-  expect_is(plot(comp_flpr, arrange = "grid"), "plot_list")
-  expect_is(plot(comp_flpr, arrange = "grid")[[1]], "ggplot")
+  expect_is(plot(comp_flpr, grid_models = TRUE), "gtable")
 
-  expect_is(plot(comp_flmdls, arrange = "grid"), "plot_list")
-  expect_is(plot(comp_flmdls, arrange = "grid")[[1]], "ggplot")
+  expect_is(plot(comp_flmdls, grid_models = TRUE), "gtable")
+})
+
+test_that("rubin_full cross-validation works", {
+  # Run it first with test data that includes baseline, this will gen a message:
+  bg <- expect_warning(expect_message(
+    baggr(subset(schools_ipd, group != "School A"), iter = 20, refresh = 0,
+          test_data = subset(schools_ipd, group == "School A")),
+    "Baselines for all these groups"
+  )
+  )
+
+  # Now repeat without bsl data
+  bg <- expect_warning(baggr(subset(schools_ipd, group != "School A"), iter = 20, refresh = 0,
+          test_data = subset(schools_ipd, group == "School A" & treatment == 1)))
+  expect_is(bg, "baggr")
+  expect_gt(bg$mean_lpd, 0)
 })